+++ src/app/core/validators/index.ts
@@ -1,0 +1,3 @@
+export * from './custom.validators';
+export * from './reg-exp.service';
+export * from './email.directive';