+++ src/app/core/validators/reg-exp.service.ts
@@ -1,0 +1,4 @@
+export class RegExpService {
+  static onlyLetters = /^\S+$/g;
+  static email = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
+}