+++ src/app/orders/components/index.ts
@@ -1,7 +1,10 @@
 import { OrderListComponent } from './order-list';
+import { ProcessOrderComponent } from './process-order';
 
 export const components = [
-  OrderListComponent
+  OrderListComponent,
+  ProcessOrderComponent,
 ];
 
 export * from './order-list';
+export * from './process-order';