+++ src/app/orders/components/process-order/index.ts
@@ -1,0 +1,1 @@
+export * from './process-order.component';