+++ src/app/shared/components/input-errors/input-errors.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+import { FormControl } from '@angular/forms';
+
+@Component({
+  selector: 'sh-input-errors',
+  templateUrl: './input-errors.component.html',
+  styleUrls: ['./input-errors.component.scss']
+})
+
+export class InputErrorsComponent {
+  @Input() public control: FormControl;
+}
+